Fixed pause menu bug
The Start() function of the GameController needed to be changed so the
scene could determine what children of the pause menu canvas should be
active in the scene. When this is not done the Unity editor will default
to setting all children of the canvas active after pulling from git,
resulting in a mess.